pyproject.toml
[project]
name = "insonia-v2"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "fastapi-users[sqlalchemy]>=12.1.3",
    "psycopg>=3.3.3",
    "pydantic>=2.13.3",
    "python-decouple>=3.8",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.49",
    "strawberry-graphql[fastapi]>=0.314.3",
    "uvicorn[standard]>=0.46.0",
]
[tool.pylint.messages_control]
disable = ["no-member"]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.3.0",
]
[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"
